+++ recycler-decorator/sample/src/main/java/ru/surfstudio/android/recycler/decorator/sample/EasyAdapterDecoratorActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
-import kotlinx.android.synthetic.main.activity_linear_decoration.*
+import kotlinx.android.synthetic.main.activity_recycler.*
 import ru.surfstudio.android.easyadapter.EasyAdapter
 import ru.surfstudio.android.easyadapter.ItemList
 import ru.surfstudio.android.recycler.decorator.Builder
@@ -24,7 +24,7 @@ class EasyAdapterDecoratorActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_linear_decoration)
+        setContentView(R.layout.activity_recycler)
         init()
     }