+++ recycler-decorator/sample/src/main/java/ru/surfstudio/android/recycler/decorator/sample/LinearDecoratorActivity.kt
@@ -6,7 +6,7 @@ import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import ru.surfstudio.android.recycler.decorator.sample.controllers.Controller
-import kotlinx.android.synthetic.main.activity_linear_decoration.*
+import kotlinx.android.synthetic.main.activity_recycler.*
 import ru.surfstudio.android.easyadapter.EasyAdapter
 import ru.surfstudio.android.easyadapter.ItemList
 import ru.surfstudio.android.recycler.decorator.Builder
@@ -23,7 +23,7 @@ class LinearDecoratorActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_linear_decoration)
+        setContentView(R.layout.activity_recycler)
         init()
     }