+++ recycler-decorator/sample/src/main/java/ru/surfstudio/android/recycler/decorator/sample/easydecor/EasyItemDecor.kt
@@ -2,10 +2,12 @@ package ru.surfstudio.android.recycler.decorator.sample.easydecor
 
 import android.graphics.Canvas
 import android.graphics.Rect
+import android.util.Log
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import ru.surfstudio.android.easyadapter.EasyAdapter
 import ru.surfstudio.android.easyadapter.item.BaseItem
+import ru.surfstudio.android.easyadapter.item.NoDataItem
 import ru.surfstudio.android.recycler.decorator.Builder
 import ru.surfstudio.android.recycler.decorator.base.OffsetDecor
 import ru.surfstudio.android.recycler.decorator.base.ViewHolderDecor
@@ -21,7 +23,11 @@ class BaseItemControllerDecoration<I : BaseItem<out RecyclerView.ViewHolder>>(
 
         val adapter = recyclerView.adapter as EasyAdapter
 
-        val baseItem = adapter.getItem(itemPosition) as I
+        val baseItem = adapter.getItem(itemPosition) as? I
+
+        if (adapter.isFirstInvisibleItemEnabled && baseItem is NoDataItem<*> && itemPosition == 0) {
+            return
+        }
 
         baseViewHolderDecor.draw(canvas, view, recyclerView, state, baseItem)
     }
@@ -35,9 +41,15 @@ class BaseItemControllerOffset<I : BaseItem<out RecyclerView.ViewHolder>>(
     override fun getItemOffsets(outRect: Rect, view: View, recyclerView: RecyclerView, state: RecyclerView.State) {
         val itemPosition = recyclerView.getChildAdapterPosition(view)
 
+//        Log.d("BaseItemControllerOffset","Call draw for position: $itemPosition")
+
         val adapter = recyclerView.adapter as EasyAdapter
 
-        val baseItem = adapter.getItem(itemPosition) as I
+        val baseItem = adapter.getItem(itemPosition) as? I
+
+        if (adapter.isFirstInvisibleItemEnabled && baseItem is NoDataItem<*> && itemPosition == 0) {
+            return
+        }
 
         baseViewHolderOffset.getItemOffsets(outRect, view, recyclerView, state, baseItem)
     }
@@ -48,7 +60,7 @@ interface BaseViewHolderDecor<I : BaseItem<out RecyclerView.ViewHolder>> {
              view: View,
              recyclerView: RecyclerView,
              state: RecyclerView.State,
-             baseItem: I)
+             baseItem: I?)
 }
 
 interface BaseViewHolderOffset<I : BaseItem<out RecyclerView.ViewHolder>> {
@@ -56,7 +68,7 @@ interface BaseViewHolderOffset<I : BaseItem<out RecyclerView.ViewHolder>> {
                        view: View,
                        recyclerView: RecyclerView,
                        state: RecyclerView.State,
-                       baseItem: I)
+                       baseItem: I?)
 }