+++ easyadapter/RELEASE_NOTES.md
@@ -9,6 +9,7 @@
 * ANDDEP-812 - Fixed crash after enabling infinite scroll, fixed scroll position saving mechanism
 * ANDDEP-829 Fixed synchronous DiffResult dispatching in DefaultDiffer.
 * ANDDEP-785 Added method getItem(), for get BaseItem by position
+* ANDDEP-785 Added method getter, for *firstInvisibleItemEnabled* property
 ## 1.0.0
 ##### Easyadapter
 * ANDDEP-270 Added support for asynchronous view inflate in ViewHolder